# go/api_store.go
/*
 * OpenAPI Petstore
 *
 * This is a sample server Petstore server. For this sample, you can use the api key `special-key` to test the authorization filters.
 *
 * API version: 1.0.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package petstoreserver

import (
	"net/http"

	"github.com/gin-gonic/gin"

	storehttpmapper "github.com/GIT_USER_ID/GIT_REPO_ID/internal/store/adapters/http/mapper"
	storeapp "github.com/GIT_USER_ID/GIT_REPO_ID/internal/store/application"
	storeports "github.com/GIT_USER_ID/GIT_REPO_ID/internal/store/ports"
)

// StoreAPI implements the store/order OpenAPI operations.
type StoreAPI struct {
	service *storeapp.Service
}

// NewStoreAPI wires the application service.
func NewStoreAPI(service *storeapp.Service) StoreAPI {
	return StoreAPI{service: service}
}

func toTransportOrder(model Order) storehttpmapper.Order {
	return storehttpmapper.Order{
		ID:       model.Id,
		PetID:    model.PetId,
		Quantity: model.Quantity,
		ShipDate: model.ShipDate,
		Status:   model.Status,
		Complete: model.Complete,
	}
}

func fromTransportOrder(order storehttpmapper.Order) Order {
	return Order{
		Id:       order.ID,
		PetId:    order.PetID,
		Quantity: order.Quantity,
		ShipDate: order.ShipDate,
		Status:   order.Status,
		Complete: order.Complete,
	}
}

// Delete /v2/store/order/:orderId
// Delete purchase order by ID
func (api *StoreAPI) DeleteOrder(c *gin.Context) {
	id, ok := parseIDParam(c, "orderId")
	if !ok {
		return
	}
	if err := api.service.DeleteOrder(c.Request.Context(), id); err != nil {
		respondStoreError(c, err)
		return
	}
	c.Status(http.StatusOK)
}

// Get /v2/store/inventory
// Returns pet inventories by status
func (api *StoreAPI) GetInventory(c *gin.Context) {
	inv, err := api.service.Inventory(c.Request.Context())
	if err != nil {
		respondStoreError(c, err)
		return
	}
	c.JSON(http.StatusOK, inv)
}

// Get /v2/store/order/:orderId
// Find purchase order by ID
func (api *StoreAPI) GetOrderById(c *gin.Context) {
	id, ok := parseIDParam(c, "orderId")
	if !ok {
		return
	}
	order, err := api.service.GetOrderByID(c.Request.Context(), id)
	if err != nil {
		respondStoreError(c, err)
		return
	}
	c.JSON(http.StatusOK, fromTransportOrder(storehttpmapper.FromDomainOrder(order)))
}

// Post /v2/store/order
// Place an order for a pet
func (api *StoreAPI) PlaceOrder(c *gin.Context) {
	var payload Order
	if err := c.ShouldBindJSON(&payload); err != nil {
		respondError(c, http.StatusBadRequest, err)
		return
	}
	order := storehttpmapper.ToDomainOrder(toTransportOrder(payload))
	saved, err := api.service.PlaceOrder(c.Request.Context(), order)
	if err != nil {
		respondStoreError(c, err)
		return
	}
	c.JSON(http.StatusOK, fromTransportOrder(storehttpmapper.FromDomainOrder(saved)))
}

func respondStoreError(c *gin.Context, err error) {
	if err == nil {
		return
	}
	if err == storeports.ErrNotFound {
		respondError(c, http.StatusNotFound, err)
		return
	}
	respondError(c, http.StatusInternalServerError, err)
}