# go/model_pet.go
/*
 * OpenAPI Petstore
 *
 * This is a sample server Petstore server. For this sample, you can use the api key `special-key` to test the authorization filters.
 *
 * API version: 1.0.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package petstoreserver

import "time"

// Pet - A pet for sale in the pet store
type Pet struct {
	Id int64 `json:"id,omitempty"`

	Category Category `json:"category,omitempty"`

	Name string `json:"name"`

	PhotoUrls []string `json:"photoUrls"`

	Tags []Tag `json:"tags,omitempty"`

	// pet status in the store
	// Deprecated
	Status            string                `json:"status,omitempty"`
	HairLengthCm      *float64              `json:"hairLengthCm,omitempty"`
	ExternalReference *PetExternalReference `json:"externalReference,omitempty"`
	CreatedAt         time.Time             `json:"createdAt,omitempty"`
	UpdatedAt         time.Time             `json:"updatedAt,omitempty"`
}

// PetExternalReference captures how the local pet maps to a remote provider record.
type PetExternalReference struct {
	Provider   string            `json:"provider,omitempty"`
	ExternalId string            `json:"externalId,omitempty"`
	Attributes map[string]string `json:"attributes,omitempty"`
}

// GroomingOperation carries transient hair-measurement data for the grooming endpoint.
type GroomingOperation struct {
	InitialHairLengthCm float64 `json:"initialHairLengthCm"`
	TrimByCm            float64 `json:"trimByCm"`
}